Replace xml.dom with xml.etree.ElementTree
Rather than using the xml.dom implementation, which is more fiddly
(as it inserts extra nodes for things like text attributes), and having
to roll our own XPath implementation, use the built in xml.etree.ElementTree,
which is more Pythonic and has basic XPath.

(In future, for performance, we could also allow switching out to lxml if
it's available instead, with xml.etree.ElementTree as a fallback).