Initialize KalturaClient.ks to None
It is customary in Python to initialize all class members in the `.__init__()` method.

Sometimes `KalturaClient.setKs()` fails to set `KalturaClient.ks` without raising an exception.  In our code, `client.setKs()` is called and no exception is raised but subsequently accessing `client.ks` raises an `AttributeError`.  It would be much more intuitive to test `if client.ks:` at anytime.